Add support for setting profiles to manufacturer defaults
The G.Skill MX780 seems to have a command that allows us to request that
the mouse completely wipe a profile and replace it with the
manufacturer's defaults. The defaults are stored on the mouse itself.

At the moment none of the drivers support this (but the G.Skill driver
will).

This is pretty useful, especially if you're working on a driver and you
accidentally replace a good profile with one that's bogus :).