gskill-mx780: Add support for reading/writing buttons
This adds support for basic button bindings on the G.Skill Ripjaws
MX780. There's a couple of things this leaves support out for,
specifically the ability to switch to a specific resolution/profile (as
opposed to cycling through the available profiles). However these things
are going to require extra support in libratbag.